refactor - get Metadata + footer on assets
