+++ client/src/app/(validator)/data-validation/assets/content.tsx
@@ -1,47 +1,19 @@
 import React from "react";
 import { useSearchParams } from "next/navigation";
 
-import { type Metadata } from "~/lib/db/types";
+import type { MetatdataJSON } from "~/lib/types";
 
 import TableView from "./table";
 import GridView from "./grid";
-import { bind } from "~/lib/bind-number";
-import { useActiveProject } from "~/providers/active-project.provider";
-import { useRxData } from "rxdb-hooks";
-import Loader from "~/components/loader";
-import { chunk } from "~/lib/chunk";
+import { getPageFromParams } from "~/lib/get-page-from-param";
 
-export default function Content() {
+export default function Content({ metadata }: { metadata: MetatdataJSON[] }) {
   const searchParams = useSearchParams();
   const view = searchParams.get("view") ?? "table"; // Default to : table
-  const activeProject = useActiveProject();
-  const { result, isFetching } = useRxData<Metadata>("metadata", (collection) =>
-    collection.findByIds([activeProject?.metadataId ?? ""]),
-  );
-
-  if (isFetching)
-    return (
-      <div className="flex items-center justify-center">
-        <Loader />
-      </div>
-    );
-
-  const metadata = result[0]?.data;
-
-  if (!activeProject || !metadata) return <div>No data found.</div>;
-
-  const pagedMetadata = chunk(metadata, 10);
   const currentPage = searchParams.get("page");
-  const page = getPageFromParams(currentPage, pagedMetadata.length);
+  const page = getPageFromParams(currentPage, metadata.length);
 
-  if (view === "table")
-    return <TableView metadata={pagedMetadata} page={page} />;
+  if (view === "table") return <TableView metadata={metadata} page={page} />;
 
-  return <GridView metadata={pagedMetadata} page={page} />;
+  return <GridView metadata={metadata} page={page} />;
 }
-
-const getPageFromParams = (param: string | null, max: number) => {
-  if (!param) return 1;
-  if (isNaN(+param)) return 1;
-  return bind(1, max, +param);
-};
