+++ client/src/app/(validator)/data-validation/assets/grid/index.tsx
@@ -1,7 +1,6 @@
 import { type Metadata } from "~/lib/db/types";
 
 import Card from "./card";
-import Footer from "../footer";
 
 type GridViewProps = {
   metadata: Metadata["data"][];
@@ -10,13 +9,10 @@ type GridViewProps = {
 
 export default function GridView({ metadata, page }: GridViewProps) {
   return (
-    <>
-      <div className="grid grid-flow-col grid-rows-2 gap-4 px-4">
-        {metadata[page - 1]?.map((meta) => (
-          <Card key={meta.assetName} asset={meta} />
-        ))}
-      </div>
-      <Footer page={page} lastPage={metadata.length} />
-    </>
+    <div className="grid grid-flow-col grid-rows-2 gap-4 px-4">
+      {metadata[page - 1]?.map((meta) => (
+        <Card key={meta.assetName} asset={meta} />
+      ))}
+    </div>
   );
 }
